Fixed issue with losing JVM dying too fast
I'm not entirely sure what the problem is/was, but now after the JVM is
loaded, the SCM thread now attaches to the JVM and keeps a reference to
the service object.

Some of our services seemed to work fine without this cross-thread
jobject reference, but for whatever reason some of them would have the
jvm shutdown before their startup threads could fully instantiate.

This is likely not a great solution, as I believe the service requires
calling System.exit() in order to close, regardless of whether all
remaining JVM threads are deamons or not. I think JVM's are supposed to
auto-shutdown when all non-daemon threads are closed.